+++ DemoAppSwiftUI/CustomComposerAttachmentView.swift
@@ -231,6 +231,28 @@ struct CustomContactAttachmentComposerPreview: View {
     }
 }
 
+struct DiscardAttachmentButton: View {
+    var attachmentIdentifier: String
+    var onDiscard: (String) -> Void
+
+    public init(attachmentIdentifier: String, onDiscard: @escaping (String) -> Void) {
+        self.attachmentIdentifier = attachmentIdentifier
+        self.onDiscard = onDiscard
+    }
+
+    public var body: some View {
+        TopRightView {
+            Button(action: {
+                withAnimation {
+                    onDiscard(attachmentIdentifier)
+                }
+            }, label: {
+                DiscardButtonView()
+            })
+        }
+    }
+}
+
 struct CustomContactAttachmentPreview: View {
     @Injected(\.fonts) var fonts
     @Injected(\.colors) var colors
