--- Sources/StreamChatSwiftUI/ChatMessageList/Polls/PollsHelpers.swift
@@ -1,45 +1,0 @@
-//
-// Copyright © 2026 Stream.io Inc. All rights reserved.
-//
-
-import StreamChat
-import StreamChatCommonUI
-import SwiftUI
-
-private struct PollsBackgroundModifier: ViewModifier {
-    @Injected(\.colors) var colors
-    
-    func body(content: Content) -> some View {
-        content
-            .padding()
-            .background(Color(colors.background1))
-            .cornerRadius(16)
-    }
-}
-
-extension View {
-    func withPollsBackground() -> some View {
-        modifier(PollsBackgroundModifier())
-    }
-}
-
-struct PollDateIndicatorView: View {
-    @Injected(\.fonts) var fonts
-    @Injected(\.utils) var utils
-    @Injected(\.colors) var colors
-    
-    let date: Date
-    
-    var body: some View {
-        Text(text)
-            .font(fonts.subheadline)
-            .foregroundColor(Color(colors.textLowEmphasis))
-    }
-    
-    var text: String {
-        let formatter = utils.pollsDateFormatter
-        let voteDay = formatter.formatDay(date)
-        let voteTime = formatter.formatTime(date)
-        return L10n.Message.Polls.Date.dayAtTime(voteDay, voteTime)
-    }
-}
