+++ Sources/StreamChatSwiftUI/Utils/Modifiers.swift
@@ -4,29 +4,6 @@
 
 import SwiftUI
 
-/// Modifier for adding shadow and corner radius to a view.
-struct ShadowViewModifier: ViewModifier {
-    @Injected(\.colors) private var colors
-
-    var backgroundColor: UIColor = .systemBackground
-    var cornerRadius: CGFloat = 16
-    var firstRadius: CGFloat = 10
-    var firstY: CGFloat = 12
-    
-    func body(content: Content) -> some View {
-        content.background(Color(backgroundColor))
-            .cornerRadius(cornerRadius)
-            .modifier(ShadowModifier(firstRadius: firstRadius, firstY: firstY))
-            .overlay(
-                RoundedRectangle(cornerRadius: cornerRadius)
-                    .stroke(
-                        Color(colors.innerBorder),
-                        lineWidth: 0.5
-                    )
-            )
-    }
-}
-
 /// Modifier for adding shadow to a view.
 public struct ShadowModifier: ViewModifier {
     public init(
